requirements.txt
qdrant-client==1.7.0
transformers==4.36.2
torch==2.1.2
python-dotenv==1.0.0
pytest==7.4.4
deepseek-ai==0.1.0  # DeepSeek Python SDK 
requests==2.31.0 